switch to keying by full port name instead of choosing a canonical symbolid
